optimized pop_front()
logically thinking through the process of the pop_front() function, no branch is actually necessary, as the else branch covers everything. 

notice: if we decided to keep the branch, we should have made the if statement if( head != tail), since (head == tail) only occurs once, it should actually be the else.