created non recursive replacement for bottleneck function get_code()
Because the recursion does not know which direction to travel it is very wasteful. By creating a map of the codes by doing one recursive search it greatly reduces the time it takes to complete. Getting an item from a code does not have this problem because the code provides the direct path to the item.